Update README.md
Mohammed Alzabyedi
Project Proposal
T5 Bootcamp







Questions:


What countries in the database are more likely to have a severe weather that could impact the spread of Covid-19?

What countries in the database are less likely to have a severe weather that could impact the spread of Covid-19?

From this project my findings could help the CDC and all the healthcare providers in the countries am using to know the likelihood of Covid-19 spread among their citizens.

Data Description:

I am going to use a weather forecast dataset for more than one country using Kaggle datasets. My sample is a dataset of some countries in which Covid-19 could spread.


Tools:

For this project am going to follow the given guidelines in Github. Also, I will be using some tools for my approach including Pandas, Numpy and Seaborn for my Visualization.